requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
openai>=1.12.0
pydantic>=2.0.0
python-dotenv>=1.0.0
slowapi>=0.1.9
psycopg2-binary>=2.9.0
httpx>=0.27.0
